+++ torchvision/models/detection/roi_heads.py
@@ -4,8 +4,7 @@ import torch
 import torch.nn.functional as F
 import torchvision
 from torch import nn, Tensor
-from torchvision.ops import boxes as box_ops
-from torchvision.ops import roi_align
+from torchvision.ops import boxes as box_ops, roi_align
 
 from . import _utils as det_utils
