+++ torchvision/prototype/models/__init__.py
@@ -1,4 +1,3 @@
-from . import detection
 from . import optical_flow
 from . import segmentation
 from . import video
