--- torchvision/prototype/models/detection/__init__.py
@@ -1,7 +1,0 @@
-from .faster_rcnn import *
-from .fcos import *
-from .keypoint_rcnn import *
-from .mask_rcnn import *
-from .retinanet import *
-from .ssd import *
-from .ssdlite import *
